+++ src/frontend/apps/desk/src/features/mail-domains/mailboxes/components/panel/MailBoxesListView.tsx
@@ -97,6 +97,7 @@ export function MailBoxesListView({
             {
               field: 'email',
               headerName: `${t('Address')} • ${filteredMailboxes.length}`,
+              renderCell: ({ row }) => <Text>{row.email}</Text>,
             },
             {
               field: 'name',